requirements.txt
Django==1.8.3
wsgiref==0.1.2
django-templated-email